Backend/requirements.txt
flask
appwrite
python-dotenv
requests
werkzeug
flask-cors
razorpay
